--- src/utils/myAccountData.ts
@@ -1,7 +1,0 @@
-export const notificationPreferenceMap = {
-  z: "Email",
-  a: "Print",
-  p: "Phone",
-  m: "Mobile",
-  "-": null,
-}
